vinylla_server/requirements.txt
asgiref==3.8.1
attrs==24.1.0
black==24.8.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Django>=4
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
flake8==7.1.1
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
pycodestyle==2.12.1
pyflakes==3.2.0
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.1
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.2
